:sparkles: Added Vim Rails config and a-z'd config
A gripe I always had with vim-rails is that :Ttest never worked for me.
In my current project we have a lot of files under services for which
the auto load wasn't working either.

So I finally got fed up and read the documentation on how to configure
it properly. There are still a couple of things I didn't quite get
working (like load related tests), but this honestly works way better
than before.